タブ管理機能のハンドラを追加・整理
