+++ .oxlintrc.json
@@ -74,7 +74,7 @@
     "eslint/max-lines-per-function": [
       "error",
       {
-        "max": 1000,
+        "max": 500,
         "skipBlankLines": true,
         "skipComments": true
       }
