+++ src/contexts/saved-tabs/presentation/hooks/useProjectCategoryHandlers.ts
@@ -1,0 +1,302 @@
+import { useCallback, useEffect } from 'react'
+import type { Dispatch, SetStateAction } from 'react'
+import { toast } from 'sonner'
+
+import { customProjectIdToString } from '@/contexts/saved-tabs/domain/value-objects/CustomProjectId'
+import type { CustomProject, ViewMode } from '@/types/storage'
+
+import {
+  toRawStorageCustomProject,
+} from './projectManagementDefaults'
+import type { ProjectManagementRefs } from './useProjectManagementRefs'
+
+interface ProjectCategoryHandlerDeps {
+  refs: ProjectManagementRefs
+  setCustomProjects: Dispatch<SetStateAction<CustomProject[]>>
+  setViewMode: Dispatch<SetStateAction<ViewMode>>
+  initialViewMode: ViewMode | undefined
+  t: (key: string, fallback?: string, values?: Record<string, string>) => string
+}
+
+const useProjectCategoryHandlers = ({
+  refs,
+  setCustomProjects,
+  setViewMode,
+  initialViewMode,
+  t,
+}: ProjectCategoryHandlerDeps) => {
+  const handleAddCategory = useCallback(
+    async (projectId: string, categoryName: string): Promise<void> => {
+      try {
+        await refs.addCategoryToCustomProjectUseCaseRef.current({
+          categoryName,
+          projectId,
+        })
+        setCustomProjects((prev) =>
+          prev.map((p) => {
+            if (p.id !== projectId) {
+              return p
+            }
+            if (p.categories.includes(categoryName)) {
+              return p
+            }
+
+            const updatedCategories = [...p.categories, categoryName]
+            const baseCategoryOrder = p.categoryOrder ?? p.categories
+            return {
+              ...p,
+              categories: updatedCategories,
+              categoryOrder: baseCategoryOrder.includes(categoryName)
+                ? baseCategoryOrder
+                : [...baseCategoryOrder, categoryName],
+              updatedAt: Date.now(),
+            }
+          }),
+        )
+        toast.success(
+          t('savedTabs.projectCategory.added', undefined, {
+            name: categoryName,
+          }),
+        )
+      } catch (error) {
+        console.error('カテゴリ追加エラー:', error)
+        toast.error(t('savedTabs.subCategory.createError'))
+      }
+    },
+    [refs.addCategoryToCustomProjectUseCaseRef, setCustomProjects, t],
+  )
+
+  const handleDeleteProjectCategory = useCallback(
+    async (projectId: string, categoryName: string): Promise<void> => {
+      try {
+        await refs.removeCategoryFromCustomProjectUseCaseRef.current({
+          categoryName,
+          projectId,
+        })
+        const updatedRaws = await refs.getCustomProjectRawsQueryRef.current()
+        setCustomProjects(updatedRaws.map(toRawStorageCustomProject))
+        toast.success(
+          t('savedTabs.projectCategory.deleted', undefined, {
+            name: categoryName,
+          }),
+        )
+      } catch (error) {
+        console.error('カテゴリ削除エラー:', error)
+        toast.error(t('savedTabs.subCategory.deleteError'))
+      }
+    },
+    [refs.removeCategoryFromCustomProjectUseCaseRef, refs.getCustomProjectRawsQueryRef, setCustomProjects, t],
+  )
+
+  const handleSetUrlCategory = useCallback(
+    async (
+      projectId: string,
+      url: string,
+      category?: string,
+    ): Promise<void> => {
+      try {
+        await refs.setCustomProjectUrlCategoryUseCaseRef.current({
+          category,
+          projectId,
+          url,
+        })
+        const updatedRaws = await refs.getCustomProjectRawsQueryRef.current()
+        setCustomProjects(updatedRaws.map(toRawStorageCustomProject))
+      } catch (error) {
+        console.error('URL分類エラー:', error)
+        toast.error(t('savedTabs.tab.moveError'))
+      }
+    },
+    [refs.setCustomProjectUrlCategoryUseCaseRef, refs.getCustomProjectRawsQueryRef, setCustomProjects, t],
+  )
+
+  const handleUpdateCategoryOrder = useCallback(
+    async (projectId: string, newOrder: string[]): Promise<void> => {
+      try {
+        console.log(`カテゴリ順序を更新: ${projectId}`, newOrder)
+        await refs.updateCustomProjectCategoryOrderUseCaseRef.current({
+          newOrder,
+          projectId,
+        })
+        setCustomProjects((prev) =>
+          prev.map((p) =>
+            p.id === projectId
+              ? {
+                  ...p,
+                  categoryOrder: newOrder,
+                  updatedAt: Date.now(),
+                }
+              : p,
+          ),
+        )
+      } catch (error) {
+        console.error('カテゴリ順序更新エラー:', error)
+        toast.error(t('savedTabs.projectCategory.orderUpdateError'))
+      }
+    },
+    [refs.updateCustomProjectCategoryOrderUseCaseRef, setCustomProjects, t],
+  )
+
+  const handleReorderUrls = useCallback(
+    async (projectId: string, urls: CustomProject['urls']): Promise<void> => {
+      try {
+        await refs.reorderCustomProjectUrlsUseCaseRef.current({
+          projectId,
+          urls,
+        })
+        setCustomProjects((prev) =>
+          prev.map((p) =>
+            p.id === projectId
+              ? {
+                  ...p,
+                  updatedAt: Date.now(),
+                  urls,
+                }
+              : p,
+          ),
+        )
+      } catch (error) {
+        console.error('URL順序更新エラー:', error)
+        toast.error(t('savedTabs.tab.orderUpdateError'))
+      }
+    },
+    [refs.reorderCustomProjectUrlsUseCaseRef, setCustomProjects, t],
+  )
+
+  const handleReorderProjects = useCallback(
+    async (newOrder: string[]): Promise<void> => {
+      try {
+        console.log('プロジェクト順序を更新:', newOrder)
+        await refs.saveCustomProjectOrderUseCaseRef.current({ newOrder })
+        setCustomProjects((prev) =>
+          prev.toSorted((a, b) => {
+            const indexA = newOrder.indexOf(a.id)
+            const indexB = newOrder.indexOf(b.id)
+            if (indexA === -1) {
+              return 1
+            }
+            if (indexB === -1) {
+              return -1
+            }
+            return indexA - indexB
+          }),
+        )
+        toast.success(t('savedTabs.projects.orderUpdated'))
+      } catch (error) {
+        console.error('プロジェクト順序更新エラー:', error)
+        toast.error(t('savedTabs.projects.orderUpdateError'))
+      }
+    },
+    [refs.saveCustomProjectOrderUseCaseRef, setCustomProjects, t],
+  )
+
+  const handleRenameCategory = useCallback(
+    async (
+      projectId: string,
+      oldCategoryName: string,
+      newCategoryName: string,
+    ): Promise<void> => {
+      try {
+        await refs.renameCustomProjectCategoryUseCaseRef.current({
+          newCategoryName,
+          oldCategoryName,
+          projectId,
+        })
+        setCustomProjects((prev) =>
+          prev.map((project) =>
+            project.id === projectId
+              ? {
+                  ...project,
+                  // eslint-disable-next-line eslint/max-nested-callbacks
+                  categories: project.categories.map((cat) =>
+                    cat === oldCategoryName ? newCategoryName : cat,
+                  ),
+                  categoryOrder: project.categoryOrder
+                    ? // eslint-disable-next-line eslint/max-nested-callbacks
+                      project.categoryOrder.map((cat) =>
+                        cat === oldCategoryName ? newCategoryName : cat,
+                      )
+                    : project.categoryOrder,
+                  // eslint-disable-next-line eslint/max-nested-callbacks
+                  urls: project.urls?.map((item) => ({
+                    ...item,
+                    category:
+                      item.category === oldCategoryName
+                        ? newCategoryName
+                        : item.category,
+                  })),
+                }
+              : project,
+          ),
+        )
+        toast.success(t('savedTabs.projectCategory.renamed'))
+      } catch (error) {
+        console.error('カテゴリ名の変更エラー:', error)
+        toast.error(t('savedTabs.subCategory.renameError'))
+      }
+    },
+    [refs.renameCustomProjectCategoryUseCaseRef, setCustomProjects, t],
+  )
+
+  useEffect(() => {
+    let isActive = true
+
+    const loadProjects = async () => {
+      try {
+        console.log(
+          '初回ロード: ビューモードとカスタムプロジェクトを取得します',
+        )
+        const mode = initialViewMode ?? 'domain'
+        setViewMode(mode)
+        console.log(`ビューモード: ${mode}`)
+
+        const [raws, order] = await Promise.all([
+          refs.getCustomProjectRawsQueryRef.current(),
+          refs.getCustomProjectOrderQueryRef.current(),
+        ])
+        const projectsAsCust = raws.map(toRawStorageCustomProject)
+        const orderKeys = order.map(customProjectIdToString)
+        const ordered =
+          orderKeys.length > 0
+            ? [
+                ...orderKeys
+                  .map((id) =>
+                    projectsAsCust.find((project) => project.id === id),
+                  )
+                  .filter(
+                    (project): project is CustomProject =>
+                      project !== undefined,
+                  ),
+                ...projectsAsCust.filter(
+                  (project) => !orderKeys.includes(project.id),
+                ),
+              ]
+            : projectsAsCust
+        console.log(`カスタムプロジェクト数: ${ordered.length}`)
+
+        if (isActive) {
+          setCustomProjects(ordered)
+        }
+        console.log('初回ロード完了')
+      } catch (error) {
+        console.error('ビューモードの読み込みエラー:', error)
+      }
+    }
+    void loadProjects()
+    return () => {
+      isActive = false
+    }
+  }, [initialViewMode, refs.getCustomProjectOrderQueryRef, refs.getCustomProjectRawsQueryRef, setCustomProjects, setViewMode])
+
+  return {
+    handleAddCategory,
+    handleDeleteProjectCategory,
+    handleRenameCategory,
+    handleReorderProjects,
+    handleReorderUrls,
+    handleSetUrlCategory,
+    handleUpdateCategoryOrder,
+  }
+}
+
+export { useProjectCategoryHandlers }
